Source commit: 9110354b1d20b2f5141620ed8859a0b6102bcf2e

+++ source/console/Console.h
@@ -77,7 +77,7 @@ struct TConsoleCommStruct
   struct TTransferEvent
   {
     unsigned char Data[20480];
-    size_t Len;
+    unsigned int Len;
     bool Error; // TRANSFERIN only
   };
